♻️refactor: 検索バーコンテナを _SearchBarContainer に抽出し MailAppBar・MailSearchAppBar で共通化
MailAppBar と _SearchAppBar（→ MailSearchAppBar に改名）で同一の Container+Row スタイルが重複していたため、
_SearchBarContainer ウィジェットに切り出して両方から参照するよう変更。
併せて _SearchAppBar を mail_app_bar.dart へ移動し MailSearchAppBar として公開することで
プライベートウィジェット間の参照を一ファイル内に収めた。
